Add --first-root option
This is useful in jhbuild when I often only want the tests from
that root, rather than also running the system tests.

Perhaps in the future we should optimize this further so that the
first directory "wins" for tests rather than being merged, e.g.  the
ostree/ tests in jhbuild win over the system root.