runner: Allow overriding test environment
In the fallout of https://bugzilla.gnome.org/show_bug.cgi?id=732184
tests exercising deprecated properties are now failing.

When running under `make check`, we can change the environment of a test
by using the TEST_ENVIRONMENT variable. The .test file that we use to
describe a test should have the same facility: a `TestEnvironment` key
that contains a list of key=value pairs that get set in the environment
prior to execution.